backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.0.0
blosc2==2.3.2
boto3==1.34.44
botocore==1.34.44
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
contourpy==1.1.0
coverage==7.3.2
cryptography==41.0.1
cssutils==2.9.0
curl-cffi==0.11.3
cycler==0.11.0
dataframe-image==0.2.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.0.0
django==4.2.2
django-cors-headers==4.4.0
django-filter==23.5
django-heroku==0.3.1
django-hosts==6.0
django-nose==1.4.7
django-rq==2.10.1
django-storages==1.13.2
django-tables2==2.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
eod==0.2.1
et-xmlfile==2.0.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.40.0
fpdf==1.7.2
fqdn==1.5.1
fredapi==0.5.1
frozendict==2.4.6
frozenlist==1.4.1
graphql-core==3.2.4
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
harfile==0.3.0
html2image==2.0.4.3
html5lib==1.1
httpcore==1.0.3
httpx==0.26.0
hypothesis==6.112.1
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.12.3
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
junit-xml==1.9
jupyter-client==8.2.0
jupyter-core==5.3.1
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.3.0
kiwisolver==1.4.4
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
msgpack==1.0.7
multidict==6.0.4
multitasking==0.0.11
nbclient==0.8.0
nbconvert==7.16.1
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.5.6
nose==1.3.7
numexpr==2.8.7
numpy==1.26.4
openpyxl==3.1.5
overrides==7.4.0
packaging==24.2
pandas==2.3.0
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
peewee==3.18.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.5.0
pip==25.1.1
pip-review==1.3.0
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==6.31.1
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pycparser==2.22
pydantic==1.10.12
pydantic-core==2.16.2
pygments==2.15.1
pyjwt==2.9.0
pyparsing==3.0.9
pyrate-limiter==3.7.0
pytest==8.3.3
pytest-subtests==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.1
pyzmq==26.4.0
qtpy==2.4.0
redis==5.0.1
referencing==0.30.2
reportlab==4.0.5
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rich==13.5.2
rpds-py==0.10.3
rq==1.15.1
s3transfer==0.10.0
schemathesis==3.36.0
scikit-learn==1.4.0
scipy==1.11.4
scout-apm==3.1.0
send2trash==1.8.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.38.5
starlette-testclient==0.4.1
stats-can==2.5.3
tables==3.9.2
tenacity==9.0.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tomli-w==1.0.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
types-python-dateutil==2.8.19.20240106
typing-extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.25.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.3.3
websockets==15.0.1
werkzeug==3.0.4
whitenoise==6.4.0
widgetsnbextension==4.0.9
wrapt==1.16.0
xlrd==2.0.1
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.63
zipp==3.15.0
